refactor: use axios instead of fetch for related configs
